refactor: restructure support/donation section into Donation component folder
